+++ packages/example-node-ts-2/src/index.ts
@@ -1,3 +1,9 @@
 import { getCwd } from 'example-node-ts';
 
 console.log(getCwd());
+
+if (process.env.foo !== 'bar') {
+  throw `Passing build time env vars didn't work!`;
+} else {
+  console.log('Passed build time env var is:', process.env.foo);
+}