+++ packages/example-node-ts-2/src/webpack.config.js
@@ -1,1 +1,3 @@
-module.exports = require('@mercateo/ws').getWebpackConfig();
+module.exports = require('@mercateo/ws').getWebpackConfig({
+  env: { foo: 'bar' }
+});